Support for more fine tuned note triggers
mus_trigger_instrument() now takes a note value including the low 8-bits
worth of finetune

+++ src/snd/music.h
@@ -161,7 +161,7 @@ typedef struct
 {
 	int channel;
 	MusInstrument *instrument;
-	Uint8 note;
+	Uint16 note;
 	int panning;
 } MusDelayedTrigger;
 
@@ -355,7 +355,7 @@ typedef struct RWops
 #endif
 
 int mus_advance_tick(void* udata);
-int mus_trigger_instrument(MusEngine* mus, int chan, MusInstrument *ins, Uint8 note, int panning);
+int mus_trigger_instrument(MusEngine* mus, int chan, MusInstrument *ins, Uint16 note, int panning);
 void mus_release(MusEngine* mus, int chan);
 void mus_init_engine(MusEngine *mus, CydEngine *cyd);
 void mus_set_song(MusEngine *mus, MusSong *song, Uint16 position);
